Complete structure refactoring.
