+++ server/src/modules/appointments/infra/typeorm/entities/Appointment.ts
@@ -7,7 +7,7 @@ import {
   ManyToOne,
   JoinColumn,
 } from 'typeorm'
-import User from './User'
+import User from '@modules/users/infra/typeorm/entities/User'
 
 @Entity('appointments')
 class Appointment {
