+++ server/src/modules/appointments/repositories/AppointmentsRepository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm'
-import Appointment from '../models/Appointment'
+import Appointment from '../infra/typeorm/entities/Appointment'
 
 @EntityRepository(Appointment)
 class AppointmentsRepository extends Repository<Appointment> {
