+++ server/src/modules/appointments/services/CreateAppointmentService.ts
@@ -1,8 +1,8 @@
 import { startOfHour } from 'date-fns'
 import { getCustomRepository } from 'typeorm'
-import Appointment from '../models/Appointment'
+import AppError from '@shared/errors/AppError'
+import Appointment from '../infra/typeorm/entities/Appointment'
 import AppointmentsRepository from '../repositories/AppointmentsRepository'
-import AppError from '../errors/AppError'
 
 interface Request {
   provider_id: string
