+++ server/src/modules/users/infra/http/middlewares/ensureAuthenticated.ts
@@ -1,7 +1,9 @@
 import { Request, Response, NextFunction } from 'express'
 import { verify } from 'jsonwebtoken'
-import authConfig from '../config/auth'
-import AppError from '../errors/AppError'
+
+import authConfig from '@config/auth'
+
+import AppError from '@shared/errors/AppError'
 
 interface TokenPayload {
   iat: number
