+++ server/src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import AuthenticateUserService from '../services/AuthenticateUserService'
+import AuthenticateUserService from '@modules/users/services/AuthenticateUserService'
 
 const sessionsRouter = Router()
