+++ server/src/modules/users/infra/http/routes/users.routes.ts
@@ -1,9 +1,9 @@
 import { Router } from 'express'
 import multer from 'multer'
-import CreateUserService from '../services/CreateUserService'
+import uploadConfig from '@config/upload'
+import CreateUserService from '@modules/users/services/CreateUserService'
+import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService'
 import ensureAuthenticated from '../middlewares/ensureAuthenticated'
-import uploadConfig from '../config/upload'
-import UpdateUserAvatarService from '../services/UpdateUserAvatarService'
 
 const usersRouter = Router()
 const upload = multer(uploadConfig)
