+++ server/src/modules/users/services/AuthenticateUserService.ts
@@ -1,9 +1,9 @@
 import { getRepository } from 'typeorm'
 import { compare } from 'bcryptjs'
 import { sign } from 'jsonwebtoken'
-import User from '../models/User'
-import authConfig from '../config/auth'
-import AppError from '../errors/AppError'
+import AppError from '@shared/errors/AppError'
+import authConfig from '@config/auth'
+import User from '../infra/typeorm/entities/User'
 
 interface Request {
   email: string
