+++ server/src/modules/users/services/CreateUserService.ts
@@ -1,7 +1,7 @@
 import { getRepository } from 'typeorm'
 import { hash } from 'bcryptjs'
-import User from '../models/User'
-import AppError from '../errors/AppError'
+import AppError from '@shared/errors/AppError'
+import User from '../infra/typeorm/entities/User'
 
 interface Request {
   name: string
