+++ server/src/modules/users/services/UpdateUserAvatarService.ts
@@ -1,9 +1,9 @@
 import { getRepository } from 'typeorm'
 import path from 'path'
 import fs from 'fs'
-import User from '../models/User'
-import uploadConfig from '../config/upload'
-import AppError from '../errors/AppError'
+import uploadConfig from '@config/upload'
+import AppError from '@shared/errors/AppError'
+import User from '../infra/typeorm/entities/User'
 
 interface Request {
   user_id: string
