+++ server/src/shared/infra/http/routes/index.ts
@@ -1,0 +1,12 @@
+import { Router } from 'express'
+import appointmentsRouter from '@modules/appointments/infra/http/routes/appointments.routes'
+import usersRouter from '@modules/users/infra/http/routes/users.routes'
+import sessionsRouter from '@modules/users/infra/http/routes/sessions.routes'
+
+const routes = Router()
+
+routes.use('/appointments', appointmentsRouter)
+routes.use('/users', usersRouter)
+routes.use('/sessions', sessionsRouter)
+
+export default routes
