+++ server/src/shared/infra/http/server.ts
@@ -4,11 +4,11 @@ import express, { Request, Response, NextFunction } from 'express'
 import cors from 'cors'
 import 'express-async-errors'
 
+import uploadConfig from '@config/upload'
+import AppError from '@shared/errors/AppError'
 import routes from './routes'
-import uploadConfig from './config/upload'
 
-import './database'
-import AppError from './errors/AppError'
+import '@shared/infra/typeorm'
 
 const app = express()
